pyproject.toml
[project]
name = "mcp-slack-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.12.3",
    "python-dotenv>=1.1.1",
]
